requirements.txt
Flask==2.0.2
chatterbot==1.0.5
chatterbot_corpus==1.2.0
pyyaml==5.1.2
